requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
google-api-python-client==2.86.0
google-auth==2.18.0
google-auth-httplib2==0.1.0
